test&chore: ChargePointRequestTest.kt, UsePointRequestTest.kt 추가 및 Constants.kt const 상수로 추가
