+++ src/main/kotlin/io/hhplus/tdd/support/constants/Constants.kt
@@ -2,6 +2,6 @@ package io.hhplus.tdd.support.constants
 
 class PointConstants {
     companion object {
-        val MAX_POINT_VALUE = 100000000L //1억 포인트
+        const val MAX_POINT_VALUE = 100000000L //1억 포인트
     }
 }
