+++ src/test/kotlin/io/hhplus/tdd/point/controller/PointControllerTest.kt
@@ -24,7 +24,6 @@ import org.springframework.web.bind.MethodArgumentNotValidException
 import org.springframework.web.bind.annotation.ControllerAdvice
 import org.springframework.web.bind.annotation.ExceptionHandler
 
-//TODO : 이렇게
 @ControllerAdvice
 class ValidationExceptionHandler {
     @ExceptionHandler(MethodArgumentNotValidException::class)
